requirements.txt
click==6.7
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
pymongo==3.4.0
PyYAML==3.12
Werkzeug==0.11.15
